Add to cart button functional
The add to cart button, when clicked, goes to processing page that inserts a row into the database consisting of the UserID of the user that's logged in, and the RecipeID of the recipe the user has chosen. RecipeID is the primary key in the table, which allows the same UserID to add several different recipes. If UserID was the primary key, the user would only be able to add one recipe to their cart.